button_security_minus_medicals

See merge request rodex-programmer/odoo12_addons_tour_travel!3458

+++ tt_accounting/models/tt_refund.py
@@ -61,6 +61,8 @@ class TtRefundLine(models.Model):
         })
 
     def set_to_confirm(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_level_4'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         # tanya boleh tidak IVAN
         self.commission_fee = 0
         self.charge_fee = 0
@@ -478,6 +480,8 @@ class TtRefund(models.Model):
         return res
 
     def confirm_refund_from_button(self):
+        if not self.env.user.has_group('tt_base.group_tt_agent_user'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         if self.state != 'draft':
             raise UserError("Cannot Confirm because state is not 'draft'.")
 
@@ -526,6 +530,8 @@ class TtRefund(models.Model):
             _logger.info('Error Create Email Queue')
 
     def confirm_refund_from_button_ho(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_3'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         if self.state != 'draft':
             raise UserError("Cannot Confirm because state is not 'draft'.")
 
@@ -596,6 +602,8 @@ class TtRefund(models.Model):
         return res
 
     def send_refund_from_button_backend(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_3'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         self.send_refund_from_button()
 
     def send_webhook_refund(self, data, action, child_id):
@@ -702,6 +710,8 @@ class TtRefund(models.Model):
         return res
 
     def validate_refund_from_button(self):
+        if not ({self.env.ref('tt_base.group_tt_agent_user').id, self.env.ref('tt_base.group_after_sales_master_level_3').id}.intersection(set(self.env.user.groups_id.ids))):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         if self.state != 'sent':
             raise UserError("Cannot Validate because state is not 'Sent'.")
 
@@ -785,6 +795,8 @@ class TtRefund(models.Model):
         return res
 
     def finalize_refund_from_button_backend(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_3'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         self.finalize_refund_from_button()
 
     def finalize_refund_from_button(self, list=[]):
@@ -890,6 +902,8 @@ class TtRefund(models.Model):
         return res
 
     def action_approve_backend(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_5'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         self.action_approve()
 
     def action_approve(self, list=[]):
@@ -1044,6 +1058,8 @@ class TtRefund(models.Model):
                 rec.set_to_done()
 
     def create_profit_loss_ledger(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_3'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         value = self.real_refund_amount - self.refund_amount
         if self.real_refund_amount != 0 and value != 0 and not self.profit_loss_created:
             debit = value >= 0 and value or 0
@@ -1075,6 +1091,8 @@ class TtRefund(models.Model):
             raise UserError(_('Profit & Loss Ledger has already been created, or HO has no Profit/Loss.'))
 
     def set_to_approve(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_level_4'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         for rec in self.refund_line_cust_ids:
             rec.sudo().unlink()
         self.write({
@@ -1085,6 +1103,8 @@ class TtRefund(models.Model):
         pass
 
     def action_payment(self):
+        if not ({self.env.ref('tt_base.group_tt_agent_user').id, self.env.ref('tt_base.group_after_sales_master_level_4').id}.intersection(set(self.env.user.groups_id.ids))):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         if self.state != 'approve':
             raise UserError("Cannot process payment to customer because state is not 'approved'.")
 
@@ -1105,6 +1125,8 @@ class TtRefund(models.Model):
         })
 
     def action_approve_cust(self):
+        if not ({self.env.ref('base.group_system').id, self.env.ref('tt_base.group_tt_agent_finance').id}.intersection(set(self.env.user.groups_id.ids))):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         if self.state != 'payment':
             raise UserError("Cannot process payment to customer because state is not 'approved'.")
         self.write({
@@ -1114,6 +1136,8 @@ class TtRefund(models.Model):
         })
 
     def action_done(self):
+        if not ({self.env.ref('tt_base.group_tt_agent_user').id, self.env.ref('tt_base.group_after_sales_master_level_4').id}.intersection(set(self.env.user.groups_id.ids))):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         state_list = ['approve_cust']
         if self.customer_parent_id.customer_parent_type_id.id == self.env.ref('tt_base.customer_type_fpo').id:
             state_list.append('approve')
@@ -1205,6 +1229,8 @@ class TtRefund(models.Model):
         })
 
     def cancel_refund_reverse_ledger(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_level_4'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         for rec in self.ledger_ids:
             if not rec.is_reversed:
                 rec.reverse_ledger()
@@ -1235,6 +1261,8 @@ class TtRefund(models.Model):
         }
 
     def toggle_is_vendor_received(self):
+        if not self.env.user.has_group('tt_base.group_after_sales_master_level_5'):
+            raise UserError('Error: Insufficient permission. Please contact your system administrator if you believe this is a mistake.')
         self.is_vendor_received = not self.is_vendor_received
 
     def print_refund_to_agent(self):
